Move make_unique to a subdir on include
This is to make it clearer that this is not a fundamental
part of the API, but rather just an auxiliary thing.
